backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
python-dotenv
sentence-transformers
pandas
torch